requirements.txt
# PROMETHEUS Requirements
# Python 3.11+

# Core LangChain
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=0.0.6
langchain-community>=0.0.20
langchain-chroma>=0.1.0

# LLM Providers
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.4.0

# Vector Store
chromadb>=0.4.22

# Data Validation
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Environment & Config
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Tokenization
tiktoken>=0.6.0

# Retry Logic
tenacity>=8.2.3

# Logging
structlog>=24.1.0

# UI - Streamlit
streamlit>=1.31.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=24.2.0
isort>=5.13.0
mypy>=1.8.0
ruff>=0.2.0
